feat : 로그인 유저와 비로그인 유저 클래스를 분리
